QmlDesigner: Fix 3D view grid color setting
Grid color was not loaded correctly on scene creation, because there
was a mismatch on how it was set. Normally, it was set as a color array,
but on scene creation it was set as a single color.
Changed it to be always set as a single color, as there is no reason
to allow array of colors for grid.

Fixes: QDS-10495
Change-Id: Idbe5d52a425eed7aed6beef9e3712b26519e5843
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>