CppEditor: improve quickfix after local renaming
Always update the local usages of the symbol under cursor even if we are
currently renaming a local symbol. Otherwise the semantic info is
classified as not up to date, which in turns limits the quick fix
operation to the ones coming from clangd.

As a next step we could make sure the semantic info is valid and up to
date before creating the code assist interface by calling a synchronous
CppUseSelectionsUpdater::update. But as this will potentially block the
ui, it will require some extra handling.

Change-Id: Icb9f523a5a21e65a2853e3b98a17c2f45c637825
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>