ScreenRecorder: Fix multiscreen handling when recording on Linux
The original approach of using Qt's "screen ID" as screen identifier for
ffmpeg does not work with the x11grab filter. x11grab requires the
identifier as defined in the DISPLAY environment variable, which it
receives now with this change.

x11grab grabs all screen contents in one big image image, so we need to
crop accordingly to get the content from the screen which the user
chose. Qt's QScreen API helps us with determining the right geometries.

Change-Id: I84a1686b151b7cb0aced09bc427114e0e9aa08bf
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Artem Sokolovskii <artem.sokolovskii@qt.io>