ImageViewer: Fix animation playback replay and loop
This fixes the issue that looped animated images do not automatically
loop and that animations can generally not be replayed from start.

It introduces the "resume" state to the tool bar icon in order to
distinguish between "playing an animation from start" and "resuming the
playback of a paused animation".

The setting cache mode QMovie::CacheAll was removed because it prevents
a looped animation from looping, and also its memory consumption (which
depends on animation dimension and frames) does not seem to be
positively outweighed by anything else.

Fixes: QTCREATORBUG-29606
Change-Id: Iaca8d93766201f5f953784be7ee6d56610e63695
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>