QmlDesigner: Flash transient scrollbars only when they are enabled
Visibility has a more complex definition in transient scrollbars since
they are transient. Also, it's meaningful when it's disabled, it
shouldn't be flashed. So here we use this ability to prevent them from
being flashed when they are disabled.

Also, the enabled property would be changed instead of the visibility
for the cases that transient scrollbars are out of the scroll area.
(TimelineWidget and TransitionEditorWidget)

Task-number: QDS-10644
Change-Id: If74d71571a216153a2d3aa09694760d7a02026ca
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>