Move most acknowledgments to qt_attributions.json
Create a qt_attributions.json and generate a documentation file from it
automatically when building documentation. The result is included as a
table in the acknowledgments page in the documentation.

Some attributions cannot (yet) move to qt_attributions.json, because
that requires the sources to be available in the project's source tree,
which is not the case for LLVM, Clazy, etc.

Remove the attributions from the README and instead point to the
documentation from there, so we get rid of that duplication.

Change-Id: I22623fe7495593ffce2e2c6c26255c27c5a8cb71
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>