QmlDesigner: Only play animations related to selected particle system
When particle mode is enabled in 3D view, only the animations related
to the selected particle system are played instead of all animations
in the scene. Timeline animations are never played, as those are
controlled with timeline controls.

Animation is considered related to selected particle system if the
animation target is either descendant or ancestor of the selected
system, or the system itself.

Fixes: QDS-10678
Change-Id: Iaaaec14f86d61c7aba2347b16bc757fc188601a0
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
(cherry picked from commit 8d4b4fd04d157774073426f7488e8376ff5f438c)
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>