Utils: Fix StringAspect ui connection
QLineEdit::textEdited is only called when the user types.
It is not triggered if the text changes due to undo/redo.
QLineEdit::textChanged is triggered for all changes.

We also have to make sure that we don't call
QLineEdit::setTextKeepingActiveCursor unnecessarily
as that would clear the undo/redo stack of the QLineEdit.

Change-Id: I19562d0804e6a34b11f19e86abd256a807e147d6
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: hjk <hjk@qt.io>