Work around issue in OptionPushButton with Qt 6.7
We want to suppress the push button's behavior to
open the menu on the `pressed` signal.
In Qt 6.7 this internal connection was changed to a
C++11 style connection to the push button's private,
so we have no direct way to remove this connection only.
Remove all connections to `pressed` (which we do not use atm)
and document the behavior correspondingly.

See qtbase/9ef16d94719b186051174d1df1d75b758f9a98c9

Change-Id: I0d50af23e80d98063da3a47713e5f59de4458819
Reviewed-by: hjk <hjk@qt.io>