ScreenRecorder: Decouple capture type from host OS
On Linux, we want to support more screen capture options than just
x11grab. Also, on Windows, we might want to support alternatives to
ddagrab.

This change replaces host OS specific control flow with a setting based
one. If there are multiple capture types available for an OS (not the
case with this initial change), a combobox in the settings page allows
the user to select one.

Change-Id: Ic57be4a47d14f09f53635e53167aa20a5d74669b
Reviewed-by: Cristian Adam <cristian.adam@qt.io>