Android: Fix build dir for Qt6/CMake + changed LIBRARY_OUTPUT_DIRECTORY
The "proper" build directory is required during different stages of the
Android application deployment. For example in order to determine the
input .json file for the androiddeployqt launch.

If the CMake target has a modified LIBRARY_OUTPUT_DIRECTORY property, an
incorrect build directory was determined. This change adds a hack which
in such case (and only for Qt6 + CMake) tries to find the right build
directory, which also contains the .json file.

Fixes: QTCREATORBUG-26479
Change-Id: I24342f696aed7a322030eb9310d3ca71bcada81c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>