QmlDesigner: Fix helper grid rendering issues
We can't use SCREEN_TEXTURE to apply fade as it doesn't contain
transparent objects. However, it turns out it's not actually necessary
if the grid itself is also transparent (opacity 0.99 is sufficient),
so SCREEN_TEXTURE use was removed.

Fixed all components of the helper grid to have same bounding boxes, as
it appears bounding box of the mesh is relevant to alpha calculations.
Now subdiv lines of the grid no longer randomly change their alpha
values when just orbiting the camera.

Orbiting the camera also caused constant property updates on the grid
material due to float math rounding errors, so round the result to
five significant digits there.

Limited the grid area to 30000 units as there is not much point
in bigger grid due to the fade and very large objects can cause other
issues.

Fixes: QDS-10821
Change-Id: I65c31e99a45fd30186dcdcb185f9978b5c1b322e
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>