SquishTests: Fix CSUP test
Function completions nowadays provide the special tooltip holding
the variations of argument lists if it is an overloaded function.
Check for it and if it is present just close it to avoid an endless
loop.
(When processing the file we use 'Cursor Down' to navigate through
the file but if the tooltip is open we just toggle between the
overloads and process the same line again and again)

Change-Id: I21742c5a460cf1537151e30c3135bebb83695703
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>