QtSupport: Store Qt versions after autodetection on startup
Restores old behavior. When QC is started with a Qt in PATH but
without existing settings it ran the autodetection and stored
the Qt versions to disk. This made it possible to restart QC later
with the same settingspath, but without having Qt in PATH.
Without this patch the Qt versions of QC only got persisted if
the settings dialog had been opened at least once and closed with
Ok (or an explicit Apply).

Change-Id: I74bd9bc0b96d5ff66e5a81200d6a269a5411a5ea
Reviewed-by: hjk <hjk@qt.io>