Nanotrace: Improve argument creation
To avoid any overhead for non activated categories we only take
references and if the the category is activated we create a string. But
std::forward_as_tuple() is not very readable.

So there is now keyValue(), dictonary() and array().

So you can now write keyValue("ids", array(foo, bar, hoo)) or
keyValue("ids", dictonary(keyValue("foo", foo), keyValue("bar", bar)))
and so on.

Change-Id: I08d268764d43fcf137b628f5d42ec37aaadd3b1a
Reviewed-by: Qt CI Patch Build Bot <ci_patchbuild_bot@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>