EditorManager: Native dialogs on macOS can't change their buttons
In this case we try to add a menu to a button.
"Native" is the default, so explicitly request a non-native message
box.

This is the message box that pops up e.g. when trying to open an
empty file foo.ui (the widget designer rejects to open it, Qt Creator provides the option to open with a different editor in the error dialog).

Task-number: QTBUG-118419
Change-Id: Iab3ee14593a55c2056303cfa16b99cea25893522
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>