WebAssembly: Avoid soft assert on toolchain auto detection
Windows only: When toolchains get restored on startup,
WebAssemblyToolChain would try to add MinGW to an environment. That
env is used for querying the compiler version. In order to find MinGW,
the MinGW toolchain is queried from ToolChainManager. But since the
ToolChainManager is not yet loaded at this moment, we get a QTC_CHEK
assert from ToolChainManager::toolChain.

This change prevents querying the ToolChainManager before it is loaded.
The compiler version can successfully be determined without MinGW in
path.

Whether we really need to query compiler versions that early is another
question outside the scope of this change.

Change-Id: I46edbb80edc58d7465e90e99f7f8381708f704a1
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>