GenericDeployStep: Fix rsync step
Don't use a copy of m_files for mkdirTask(), as m_files is
modified by the running recipe. So, when the recipe is prepared,
the m_files doesn't contain the right data, yet.

This is going to be refactored in master, as it seems the
isDeploymentNecessary() operating on mutable internals
is confusing.

Fixes: QTCREATORBUG-29609
Change-Id: I3f34584ffd9486322e8b26f95ac72b96a9306f8b
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>