Utils: Restore the prior edit-clear-locationbar icon direction
On left-to-right KDE systems, the clear icon on the right was originally
(up to QtC 11) pointing to the left. In Qt Creator 12 it started
pointing to the right.

Reproducible with KDE Plasma 5.27.

This change restores the prior mapping of "writing direction" to "clear
icon direction". It also restores the fall back to the freedesktop
standard name "clear-icon".

Amends: 67212a31ee4cc5c026b677329ace9f1fdaa3320c

Change-Id: I3d1340195a82422ac310128b9e81982e0d57a6ca
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>