NavigationWidgets: Fix the side that shortcuts open
The activationsMap logs were a specific view was opened last.

We need to clear outdated information from the navigation widgets'
(left|right) settings (if a view was opened last in the left widget,
that was written to the left widget's settings, but it was not cleared
from the right widget's settings, so if it ever was opened last on the
right side, that information stuck).

When restoring the state of the left|right widgets, we may not overwrite
the previously restored activationsMap.

Fixes: QTCREATORBUG-29770
Change-Id: I14c85c24f279208fb94707514cc4a8cba184e03c
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>