Core: Fix dialogParent for separate windows
The dialog parent is supposed to fall back to the main window, if there
is no modal window and the active window is not a popup or a splash
screen.

But testing for "popup or splash screen" does not work with
`testAnyFlags(Popup | SplashScreen)` because these flags are combined
flags - e.g. "Popup" is "Window + something", so when using
testAnyFlags(Popup), that returns true for _any_ Window, regardless of
popup or not.

Amends 60f11cf6371fd6f482d257c54a48070874dc4b73

Fixes: QTCREATORBUG-29741
Change-Id: I9e8defc6dd7193f5008decda0eda04dedc62f9df
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>