Welcome: Fix initial welcome screen layout for small window
The visibility of some Welcome screen elements depends on the size of
the Welcome screen.

When starting Qt Creator with a small initial window size, the hiding
of the side bar did not work.

This change removes the check for visibility as the main fix for
QTCREATORBUG-29800. The QTimer::singleShot is replaced with immediate
layout calculation in order to avoid another initial layout issue which
is only visible during startup.

Amends: 1c24ef6406d56c4e5e45741caf5ea023ae5f9f31

Fixes: QTCREATORBUG-29800
Change-Id: I4b058791f2ed2e85b9cbca932585b71a985445f9
Reviewed-by: Orgad Shaneh <orgads@gmail.com>