Fix crash in QmlOutline related to QmlDesigner
This is a workaround for the ugly hack that QmlDesigner forces the
EditorManager to send a currentEditorChanged signal with QmlDesigner's
custom, internal IEditor.

The currentEditorChanged signal triggered an update/re-creation of the
outline widget *for that editor*. Which is not the intended result in
the first place, but even aggrevated by the fact that QmlDesigner never
calls currentEditorChanged with the actual
EditorManager::currentEditor() again.

Work around that by only ever using the EditorManager::currentEditor()
for the outline widget.

Fixes: QTCREATORBUG-29653
Fixes: QTCREATORBUG-29702
Change-Id: I39c5bee94141a0d86c0e92879eb8b1c2b7ba64b9
Reviewed-by: <github-actions-qt-creator@cristianadam.eu>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>