Improve styling of dock title widgets
Use styled bar and set size to the "single row tool bar" size, which
makes it fit better with the rest of the layout/UI.
For example there is no "black-white-black" bars in the analyzer views
with timelines like QML Profiler, or in the Compiler Explorer. Also the
top bar(s) are nicely aligned e.g. in the default debugger layout when
the debugger is running.

Also switch to custom buttons.

Change-Id: I52f95e3fbb32718b131fde39d75a20105fb0304a
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>