Widget Designer: Adapt after dock widget redesign
Before the redesign, the dock widgets didn't have a "tool bar" at the
top, so we had the designer tool bar with the document dropdown and form
related tool buttons spanning the whole design mode width (to have a
continuous styled bar at the top of the window).

Now the dock widgets' title widget is a styled bar, so we would have the
designer tool bar, and below that the styled bar from the dock widgets,
which unnecessarily creates a big blog of styled bars.

Instead, move the designer tool bar inside the central form widget. The
title bars from the dock widgets complete the top styled bar now. This
is also more consistent with Edit mode.

Change-Id: I7d322a75a8a34120763b4dd85c44ddf674930e36
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>