ProjectExplorer: Fix sorting of file names in subpaths in wizard summary
In the wizard summary page, on Windows, the sorting of files in sub
directories is not correct. The paths got converted to userOutput before
they were tested for '/' and then assigned to a FilePath (which caused a
QTC_CHEK warning in addition).

This is reproducible with the "QDS compatible" Qt Quick application
wizard.

Fixes: QTCREATORBUG-29921
Change-Id: I45b623409831b21467e33481dcc3ca557f0f9afc
Reviewed-by: Cristian Adam <cristian.adam@qt.io>