McuSupport: Reset invalid packages to default on update
When updating kits some packages will still have the cached value of the
outdated package resulting in failure on updated. although default paths
would have given the correct path.

This patch restore package values to the default value, in addition to
tracking changes in model-names that correspond to the same kit.

Fixes: QTCREATORBUG-29855
Change-Id: Iad52df93e1e39c7f56bdb30036ac21cda8cdf670
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>